[refactor] add api directory import path
